jsontidy.py: Handle parsing errors
Currently jsontidy will die with an exception if parsing the JSON object
fails. This is less than desirable, since it's ugly, and not entirely
unexpected that from time to time invalid JSON will be provided.

This patch changes the behavior of the patch to print the error message
to stderr, and then quit with a non-zero code.