jsontidy.py: Fix input and output file being the same
Currently if the input and output files are the same, reading will fail.

The details of this is that the implementation opens the file for both
read and write at the same time, thus truncating the file before it's
read. This patch fixes that, while maintaining stdout as a write target.

The code complexity rises quite a bit, but I don't really see a better
way to solve the issue.

Fixes #3